Recompiled Linux executables.

+++ source/stitchAlignToTranscript.cpp
@@ -284,12 +284,12 @@ intScore stitchAlignToTranscript(uint rAend, uint gAend, uint rBstart, uint gBst
                 };
 
                 if (P.alignInsertionFlush.flushRight) {
-                    for (; (uint)jR<rBend-rAend-Ins; jR++ ){//flush the indel to the right as much as possible
+                    for (; jR<(int)rBend-(int)rAend-(int)Ins; jR++ ){//flush the indel to the right as much as possible
                         if (R[rAend+jR+1]!=G[gAend+jR+1] || G[gAend+jR+1]==4) {
                             break;
                         };
                     };
-                    if (jR==rBend-rAend-Ins) {//nothing left of the B-piece
+                    if (jR==(int)rBend-(int)rAend-(int)Ins) {//nothing left of the B-piece
                         return -1000009;
                     };
                 };
